test: refactor tests and relocate test utils
