+++ bc_obps/registration/tests/endpoints/test_regulated_products_endpoints.py
@@ -1,11 +1,8 @@
 import pytest
 import json
 from model_bakery import baker
-from registration.models import (
-    User,
-    RegulatedProduct,
-)
-from registration.utils import TestUtils
+from registration.models import RegulatedProduct
+from registration.tests.utils.helpers import CommonTestSetup, TestUtils
 
 pytestmark = pytest.mark.django_db
 
@@ -14,27 +11,18 @@ base_endpoint = "/api/registration/"
 content_type_json = "application/json"
 
 
-class TestRegulatedProductsEndpoint:
+class TestRegulatedProductsEndpoint(CommonTestSetup):
     endpoint = base_endpoint + "regulated_products"
-
-    def setup(self):
-        self.user: User = baker.make(User)
-        self.auth_header = {'user_guid': str(self.user.user_guid)}
-        self.auth_header_dumps = json.dumps(self.auth_header)
 
     def test_unauthorized_users_cannot_get(self):
         response = TestUtils.mock_get_with_auth_role(self, "cas_pending")
         assert response.status_code == 401
 
     def test_get_method_for_200_status(self):
-        response = TestUtils.mock_get_with_auth_role(self, "industry_user")
-        assert response.status_code == 200
-        response = TestUtils.mock_get_with_auth_role(self, "industry_user_admin")
-        assert response.status_code == 200
-        response = TestUtils.mock_get_with_auth_role(self, "cas_admin")
-        assert response.status_code == 200
-        response = TestUtils.mock_get_with_auth_role(self, "cas_analyst")
-        assert response.status_code == 200
+        roles = ["cas_analyst", "cas_admin", "industry_user", "industry_user_admin"]
+        for role in roles:
+            response = TestUtils.mock_get_with_auth_role(self, role)
+            assert response.status_code == 200
 
     def test_get_method_with_mock_data(self):
         baker.make(RegulatedProduct, _quantity=4)
