+++ bc_obps/registration/tests/endpoints/test_user_endpoints.py
@@ -1,14 +1,9 @@
-from django.forms import model_to_dict
 import pytest
 import json
-from model_bakery import baker
-from registration.models import (
-    User,
-)
 from registration.schema import UserIn
-from registration.utils import TestUtils
+from registration.tests.utils.helpers import CommonTestSetup, TestUtils
+from registration.tests.utils.bakers import app_role_baker
 import uuid
-from django.core.management import call_command
 from django.test import Client
 from registration.enums.enums import IdPs
 
@@ -21,20 +16,9 @@ content_type_json = "application/json"
 client = Client()
 
 
-@pytest.fixture(scope='function')
-def app_role_fixture():
-    # Load the fixture data into the test database
-    call_command('loaddata', 'real/appRole.json')
-
-
-class TestUserEndpoint:
+class TestUserEndpoint(CommonTestSetup):
     endpoint = base_endpoint + "user"
     endpoint_profile = endpoint + "-profile"
-
-    def setup_method(self):
-        self.user: User = baker.make(User)
-        self.auth_header = {'user_guid': str(self.user.user_guid)}
-        self.auth_header_dumps = json.dumps(self.auth_header)
 
     def test_unauthorized_users_cannot_get(self):
         # /user
@@ -100,7 +84,7 @@ class TestUserEndpoint:
         assert 'user_guid' not in content
 
     # # POST USER PROFILE BCEIDBUSINESS
-    @pytest.mark.usefixtures('app_role_fixture')
+    @pytest.mark.usefixtures('app_role_baker')
     def test_create_user_profile_bceidbusiness(self):
 
         # Arrange
@@ -141,7 +125,7 @@ class TestUserEndpoint:
         assert 'user_guid' not in content
 
     # POST USER PROFILE IDIR
-    @pytest.mark.usefixtures('app_role_fixture')
+    @pytest.mark.usefixtures('app_role_baker')
     def test_create_user_profile_idir(self):
 
         # Arrange
