+++ bc_obps/registration/tests/test_04_utils.py
@@ -14,7 +14,7 @@ from django.core.exceptions import ValidationError
 from registration.utils import extract_fields_from_dict
 from ninja.errors import HttpError
 from django.test import RequestFactory, TestCase
-from registration.utils import TestUtils
+from registration.tests.utils.helpers import TestUtils
 
 
 pytestmark = pytest.mark.django_db
